The thumbnail option now also works for external images
Changed default thumbnail size to 100px in width with auto height

Allows full size images when the user specifies either "0" or "0x0" as image size

The "current" class is now added to the list item of the currently viewed page. This is usefull for CSS styling when pagequery is used in the sidebar.